refactoring tests using parameterized

+++ tests/test_analysis.py
@@ -2,10 +2,10 @@ import os
 import unittest
 import subprocess
 import shutil
+from parameterized import parameterized
 
 tutorialDir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "../tutorial")  # Path to current folder
 has_SNOPT = os.environ.get("IMAGE") == "private"
-has_not_IPOPT = os.environ.get("COMPILERS") == "intel"
 try:
     from pyOCSM import ocsm
 except ImportError:
@@ -16,9 +16,11 @@ except ImportError:
 NPROCS = 2
 mpiCmd = ["mpirun", "-n", f"{NPROCS}"]
 gridFlag = ["--gridFile", "wing_vol_coarsen.cgns"]
-SLSQP = ["--opt", "SLSQP", "--optOptions", "{'MAXIT': 0}"]
-SNOPT = ["--opt", "SNOPT", "--optOptions", "{'Major iterations limit': 0}"]
-IPOPT = ["--opt", "IPOPT", "--optOptions", "{'max_iter': 0}"]
+optimizer = {
+    "SLSQP": ["--opt", "SLSQP", "--optOptions", "{'MAXIT': 0}"],
+    "SNOPT": ["--opt", "SNOPT", "--optOptions", "{'Major iterations limit': 0}"],
+    "IPOPT": ["--opt", "IPOPT", "--optOptions", "{'max_iter': 0}"],
+}
 
 
 class TestWingAnalysis(unittest.TestCase):
@@ -70,64 +72,29 @@ class TestWingOpt(unittest.TestCase):
         os.chdir("pyoptsparse")
         subprocess.check_call(["python", "rosenbrock.py"])
 
-    def test_wing_opt_SLSQP(self):
-        # first copy files
+    def _prepareDirAndFiles(self):
+        # copy files
         os.chdir("aero")
         shutil.copy("../ffd/ffd.xyz", ".")
         shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
+
+    @parameterized.expand(["SLSQP", "SNOPT", "IPOPT"])
+    def test_wing_opt(self, optName):
+        # Check if anything needs to be skipped based on available optimizers and modules
+        if optName == "SNOPT" and not has_SNOPT:
+            raise unittest.SkipTest("SNOPT is required for this test")
+        self._prepareDirAndFiles()
         shutil.rmtree("output", ignore_errors=True)
         cmd = ["python", "aero_opt.py"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + SLSQP)
+        subprocess.check_call(mpiCmd + cmd + gridFlag + optimizer[optName])
 
-    @unittest.skipUnless(has_SNOPT, "SNOPT is required for this test")
-    def test_wing_opt_SNOPT(self):
-        # first copy files
-        os.chdir("aero")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
-        shutil.rmtree("output", ignore_errors=True)
-        cmd = ["python", "aero_opt.py"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + SNOPT)
-
-    @unittest.skipIf(has_not_IPOPT, "temporarily skipping IPOPT tests on the intel image")
-    def test_wing_opt_IPOPT(self):
-        # first copy files
-        os.chdir("aero")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
-        shutil.rmtree("output_IPOPT", ignore_errors=True)
-        cmd = ["python", "aero_opt.py", "--output", "output_IPOPT"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + IPOPT)
-
+    @parameterized.expand(["SLSQP", "SNOPT", "IPOPT"])
     @unittest.skipIf(ocsm is None, "pyOCSM is required for this test")
-    def test_wing_opt_ESP_SLSQP(self):
-        # first copy files
-        os.chdir("aero")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
+    def test_wing_opt_ESP(self, optName):
+        self._prepareDirAndFiles()
         shutil.rmtree("output_ESP", ignore_errors=True)
         cmd = ["python", "aero_opt_esp.py", "--output", "output_ESP"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + SLSQP)
-
-    @unittest.skipUnless(has_SNOPT and ocsm is not None, "SNOPT and pyOCSM are required for this test")
-    def test_wing_opt_ESP_SNOPT(self):
-        # first copy files
-        os.chdir("aero")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
-        shutil.rmtree("output_ESP", ignore_errors=True)
-        cmd = ["python", "aero_opt_esp.py", "--output", "output_ESP"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + SNOPT)
-
-    @unittest.skipIf(ocsm is None, "pyOCSM is required for this test")
-    def test_wing_opt_ESP_IPOPT(self):
-        # first copy files
-        os.chdir("aero")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../../aero/analysis/wing_vol_coarsen.cgns", "wing_vol_coarsen.cgns")
-        shutil.rmtree("output_ESP", ignore_errors=True)
-        cmd = ["python", "aero_opt_esp.py", "--output", "output_ESP"]
-        subprocess.check_call(mpiCmd + cmd + gridFlag + IPOPT)
+        subprocess.check_call(mpiCmd + cmd + gridFlag + optimizer[optName])
 
 
 class TestAirfoilOpt(unittest.TestCase):
@@ -142,39 +109,27 @@ class TestAirfoilOpt(unittest.TestCase):
         subprocess.check_call(["python", "genFFD.py"])
         os.chdir("../")
 
-    def test_single_point_SLSQP(self):
-        os.chdir("singlepoint")
+    def _prepareDirAndFiles(self, dir):
+        os.chdir(dir)
         shutil.copy("../ffd/ffd.xyz", ".")
         shutil.copy("../mesh/n0012.cgns", ".")
         shutil.rmtree("output", ignore_errors=True)
+
+    @parameterized.expand(["SLSQP", "SNOPT"])
+    def test_single_point(self, optName):
+        if optName == "SNOPT" and not has_SNOPT:
+            raise unittest.SkipTest("SNOPT is required for this test")
+        self._prepareDirAndFiles("singlepoint")
         cmd = ["python", "airfoil_opt.py"]
-        subprocess.check_call(mpiCmd + cmd + SLSQP)
+        subprocess.check_call(mpiCmd + cmd + optimizer[optName])
 
-    def test_multipoint_SLSQP(self):
-        os.chdir("multipoint")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../mesh/n0012.cgns", ".")
-        shutil.rmtree("output", ignore_errors=True)
+    @parameterized.expand(["SLSQP", "SNOPT"])
+    def test_multipoint(self, optName):
+        if optName == "SNOPT" and not has_SNOPT:
+            raise unittest.SkipTest("SNOPT is required for this test")
+        self._prepareDirAndFiles("multipoint")
         cmd = ["python", "airfoil_multiopt.py"]
-        subprocess.check_call(mpiCmd + cmd + SLSQP)
-
-    @unittest.skipUnless(has_SNOPT, "SNOPT is required for this test")
-    def test_single_point_SNOPT(self):
-        os.chdir("singlepoint")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../mesh/n0012.cgns", ".")
-        shutil.rmtree("output", ignore_errors=True)
-        cmd = ["python", "airfoil_opt.py"]
-        subprocess.check_call(mpiCmd + cmd + SNOPT)
-
-    @unittest.skipUnless(has_SNOPT, "SNOPT is required for this test")
-    def test_multipoint_SNOPT(self):
-        os.chdir("multipoint")
-        shutil.copy("../ffd/ffd.xyz", ".")
-        shutil.copy("../mesh/n0012.cgns", ".")
-        shutil.rmtree("output", ignore_errors=True)
-        cmd = ["python", "airfoil_multiopt.py"]
-        subprocess.check_call(mpiCmd + cmd + SNOPT)
+        subprocess.check_call(mpiCmd + cmd + optimizer[optName])
 
 
 class TestOverset(unittest.TestCase):
